requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20

# PDF Processing
pdfplumber==0.11.4

# LLM
google-generativeai==0.8.4

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
